all AWS parts pulled out of base
